+++ src/Webiny/Component/OAuth2/Server/LinkedIn.php
@@ -62,29 +62,6 @@ class LinkedIn extends ServerAbstract
     }
 
     /**
-     * This method is called when user is redirected to the redirect_uri from the authorization step.
-     * Here you should process the response from OAuth2 server and extract the access token if possible.
-     * If you cannot get the access token, throw an exception.
-     *
-     * @param array $response Response from the OAuth2 server.
-     *
-     * @throws OAuth2Exception
-     * @return string Access token.
-     */
-    public function processAuthResponse($response)
-    {
-        if (!$this->isArray($response)) {
-            throw new OAuth2Exception('Invalid response while trying to get the access token.');
-        }
-
-        if (isset($response['result']['status']) && $response['result']['status'] != 200) {
-            throw new OAuth2Exception($response['result']['message']);
-        }
-
-        return $response['result']['access_token'];
-    }
-
-    /**
      * Returns the server name.
      *
      * @return string