+++ src/Webiny/Component/OAuth2/CHANGELOG.md
@@ -4,3 +4,9 @@ CHANGELOG
 1.0.0-RC1
 ---------
 * Initial version
+
+1.2.
+---------
+* removed the adoy/php-oauth2 dependency
+* created a bridge for league/oauth2-client
+* refactored OAuth2Interface